+++ local/features/src/common/header.ts
@@ -421,7 +421,7 @@ export class CommonHeader implements CommonHeaderInstance {
                   <a href="https://www.nicovideo.jp/recent" target="_blank" title="新着コメント">
                     新着コメント
                   </a>
-                  <a href="https://www.nicovideo.jp/local/features/dist/pages/mylist/index.html" target="_blank" title="mylist2">
+                  <a href="https://www.nicovideo.jp/local/features/dist/pages/mylist2/index.html" target="_blank" title="mylist2">
                     mylist2
                   </a>
                   <a href="https://www.nicovideo.jp/local/features/dist/pages/watch-history/index.html" target="_blank" title="watch-history">